Redesign Mark 3
After a brief excursion to write awesome-pass client signals are now up
and running which mean that sloppy focus should be on. Custom global key
bindings can now be made from the rc file in the master conf table. All
of the mouse commands have been moved to mouse.lua in the modules
section. Default settings only currently. Started a theme file but it
was causing awesome to freak out periodically, not sure why.