Pass Widget Integration and Screen Reconfig
Incorporated the new pass widget functionality. This meant making it so
that there is only one prompt regardless of the number of screens so
that pass-widget can have it as a parameter on construction. Deleted
some old cruft. Move the tasklist to the bottom of the awesome config
instead of being between the tag lists and widgets.